move multiprocessing get from bag
