+++ dask/tests/test_multiprocessing.py
@@ -1,0 +1,1 @@
+from dask.multiprocessing import get
